+++ x-pack/plugin/spatial/src/test/java/org/elasticsearch/xpack/spatial/util/GeoTestUtils.java
@@ -9,7 +9,6 @@ package org.elasticsearch.xpack.spatial.util;
 import org.apache.lucene.document.ShapeField;
 import org.apache.lucene.geo.GeoEncodingUtils;
 import org.apache.lucene.index.IndexableField;
-import org.apache.lucene.store.ByteBuffersDataOutput;
 import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
@@ -25,26 +24,11 @@ import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.geometry.Geometry;
 import org.elasticsearch.index.mapper.GeoShapeIndexer;
-import org.elasticsearch.xpack.spatial.index.mapper.CentroidCalculator;
-import org.elasticsearch.xpack.spatial.index.mapper.CoordinateEncoder;
-import org.elasticsearch.xpack.spatial.index.mapper.Extent;
-import org.elasticsearch.xpack.spatial.index.mapper.GeoRelation;
-import org.elasticsearch.xpack.spatial.index.mapper.TriangleTreeReader;
-import org.elasticsearch.xpack.spatial.index.mapper.TriangleTreeWriter;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-
 public class GeoTestUtils {
-
-    public static void assertRelation(GeoRelation expectedRelation, TriangleTreeReader reader, Extent extent) throws IOException {
-        GeoRelation actualRelation = reader.relateTile(extent.minX(), extent.minY(), extent.maxX(), extent.maxY());
-        assertThat(actualRelation, equalTo(expectedRelation));
-    }
 
     public static ShapeField.DecodedTriangle[] toDecodedTriangles(Geometry geometry) throws IOException {
         GeoShapeIndexer indexer = new GeoShapeIndexer(true, "test");
@@ -59,16 +43,6 @@ public class GeoTestUtils {
             ShapeField.decodeTriangle(scratch, triangles[i] = new ShapeField.DecodedTriangle());
         }
         return triangles;
-    }
-
-    public static TriangleTreeReader triangleTreeReader(Geometry geometry, CoordinateEncoder encoder) throws IOException {
-        ShapeField.DecodedTriangle[] triangles = toDecodedTriangles(geometry);
-        TriangleTreeWriter writer = new TriangleTreeWriter(Arrays.asList(triangles), encoder, new CentroidCalculator(geometry));
-        ByteBuffersDataOutput output = new ByteBuffersDataOutput();
-        writer.writeTo(output);
-        TriangleTreeReader reader = new TriangleTreeReader(encoder);
-        reader.reset(new BytesRef(output.toArrayCopy(), 0, Math.toIntExact(output.size())));
-        return reader;
     }
 
     public static double encodeDecodeLat(double lat) {
